Added Relay Class
